+++ hyoga/plot/naturalearth.py
@@ -2,13 +2,16 @@
 # GNU General Public License v3.0+ (https://www.gnu.org/licenses/gpl-3.0.txt)
 
 """
-Natural Earth plotting tools.
+This module contains functions to plot globally-available Natural Earth data
+in order to add cultural (cities, countries, etc) and physical (lakes, rivers,
+glaciers, etc) elements on plots. These use hyoga's internal shapefile plotter
+which may increase speed especially for high-definition data on small domains.
 """
 
 import matplotlib.pyplot as plt
-import shapely.geometry as sgeom
 import cartopy.crs as ccrs
 import cartopy.io.shapereader as cshp
+import hyoga.plot
 
 
 # Natural Earth internals
@@ -18,48 +21,7 @@ def add_feature(category=None, name=None, scale='10m', **kwargs):
     """Plot Natural Earth feature allowing a different color for the
     subject."""
     fname = cshp.natural_earth(resolution=scale, category=category, name=name)
-    return add_shapefile(fname, **kwargs)
-
-
-def add_shapefile(filename, ax=None, crs=None, subject=None, **kwargs):
-    """Plot shapefile geometries allowing a different color for the subject."""
-
-    # get current axes if None provided
-    ax = ax or plt.gca()
-
-    # prepare axes extent geometry
-    crs = crs or ccrs.PlateCarree()
-    axes_box = _get_extent_geometry(ax=ax, crs=crs)
-
-    # open shapefile data
-    shp = cshp.Reader(filename)
-
-    # separate subject and default kwargs
-    subject_kw = {k[8:]: kwargs[k] for k in kwargs if k.startswith('subject_')}
-    context_kw = {k: kwargs[k] for k in kwargs if not k.startswith('subject_')}
-
-    # find intersecting geometries
-    subject_geometries = []
-    context_geometries = []
-    for rec in shp.records():
-        if rec.geometry is not None and axes_box.intersects(rec.geometry):
-            name = rec.attributes.get('name', rec.attributes.get('NAME', None))
-            if subject is not None and name == subject and \
-                    rec.geometry not in subject_geometries:
-                subject_geometries.append(rec.geometry)
-            elif rec.geometry not in context_geometries:
-                context_geometries.append(rec.geometry)
-
-    # plot interseecting geometries
-    return (ax.add_geometries(context_geometries, crs, **context_kw),
-            ax.add_geometries(subject_geometries, crs, **subject_kw))
-
-
-def _get_extent_geometry(ax=None, crs=None):
-    """Return axes extent as shapely geometry."""
-    ax = ax or plt.gca()
-    west, east, south, north = ax.get_extent(crs=crs)
-    return sgeom.box(west, south, east, north)
+    return hyoga.plot.shapefile(fname, **kwargs)
 
 
 # Natural Earth cultural
@@ -115,7 +77,7 @@ def add_cities(ax=None, lang=None, include=None, exclude=None, ranks=None,
     # cities of high rank, which are numerous, and also avoids warning-like
     # messages 'posx and posy should be finite values'.
     crs = ccrs.PlateCarree()
-    axes_box = _get_extent_geometry(ax=ax, crs=crs)
+    axes_box = hyoga.plot.shapefile._get_extent_geometry(ax=ax, crs=crs)
     records = [r for r in records if axes_box.intersects(r.geometry)]
 
     # add text labels